+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * add support for text-justify and text-max-width
 * bug fix to theme filters that reference zoom level
 * breaking API change, `TilesetPreprocessor` now requires a zoom level
+
 ## 2.3.3
 
 * improve support for color expressions